Changed the type from 8 bit to 16 bit

+++ src/FakeTemperatureSensor.cpp
@@ -6,7 +6,7 @@ FakeTemperatureSensor::FakeTemperatureSensor()
   srand(time(NULL));
 }
 
-uint8_t FakeTemperatureSensor::temperature()
+uint16_t FakeTemperatureSensor::temperature()
 {
     return 17 + (rand() % 10);
 }
